Existujici stripy se znovu nestahuji

+++ dilbert.sh
@@ -11,9 +11,11 @@ fi
 i=0
 while [ "$date_current" != "$date_end" ]; do
 	date_current=$(date --date="$(date --date=${date_start} +%F) + $i day" +"%F")
-	link=$(wget -q -O - ${url}/${date_current}/ | grep '.strip.zoom.gif' | sed 's/^\(.*\) src="//;s/" title=\(.*\)$//;')
-	link="${url}${link}"
-	wget -nv -O "${date_current}.gif" "$link"
+	if [ ! -f "${date_current}.gif" ]; then
+		link=$(wget -q -O - ${url}/${date_current}/ | grep '.strip.zoom.gif' | sed 's/^\(.*\) src="//;s/" title=\(.*\)$//;')
+		link="${url}${link}"
+		wget -nv -O "${date_current}.gif" "$link"
+	fi
 	i=$(expr $i + 1)
 done
